+++ dotnet/DataSources/ProductReviewRepository.cs
@@ -258,22 +258,54 @@
 
         public async Task<bool> ValidateKeyAndToken(string key, string token, string baseUrl)
         {
-            bool validated = false;
+            bool keyAndTokenValidated = false;
+            bool keyHasAccess = false;
+
+            if (key != null && token != null)
+            {
+                ValidateKeyAndToken validateKeyAndToken = new ValidateKeyAndToken
+                {
+                    AppKey = key,
+                    AppToken = token
+                };
+                var jsonSerializedKeyAndToken = JsonConvert.SerializeObject(validateKeyAndToken);
+
+                var vtexIdRequest = new HttpRequestMessage
+                {
+                    Method = HttpMethod.Post,
+                    RequestUri = new Uri($"http://{this._httpContextAccessor.HttpContext.Request.Headers[VTEX_ACCOUNT_HEADER_NAME]}.{ENVIRONMENT}.com.br/api/vtexid/apptoken/login"),
+                    Content = new StringContent(jsonSerializedKeyAndToken, Encoding.UTF8, APPLICATION_JSON)
+                };
+
+                try
+                {
+                    var client = _clientFactory.CreateClient();
+                    var response = await client.SendAsync(vtexIdRequest);
+                    string responseContent = await response.Content.ReadAsStringAsync();
+                    _context.Vtex.Logger.Info("ValidateKeyAndToken", null, $"[{response.StatusCode}]");
+                    if (response.IsSuccessStatusCode)
+                    {
+                        var validatedKeyAndToken = JsonConvert.DeserializeObject<ValidatedKeyAndToken>(responseContent);
+                        keyAndTokenValidated = validatedKeyAndToken.AuthStatus.Equals("Success");
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _context.Vtex.Logger.Error("ValidateKeyAndToken", null, $"Error validating key and token '{key}'", ex);
+                }
+            }
 
             var request = new HttpRequestMessage
             {
                 Method = HttpMethod.Get,
-                RequestUri = new Uri($"http://licensemanager.vtexcommercestable.com.br/api/license-manager/pvt/accounts/hosts/{baseUrl}")
+                RequestUri = new Uri($"http://licensemanager.vtexcommercestable.com.br/api/license-manager/pvt/accounts/{this._httpContextAccessor.HttpContext.Request.Headers[VTEX_ACCOUNT_HEADER_NAME]}/logins/{key}/granted")
             };
 
-            if (key != null)
-            {
-                request.Headers.Add(HEADER_VTEX_APP_KEY, key);
-            }
+            string authToken = this._httpContextAccessor.HttpContext.Request.Headers[HEADER_VTEX_CREDENTIAL];
 
-            if (token != null)
+            if (authToken != null)
             {
-                request.Headers.Add(HEADER_VTEX_APP_TOKEN, token);
+                request.Headers.Add(AUTHORIZATION_HEADER_NAME, authToken);
             }
 
             try
@@ -281,15 +313,15 @@
                 var client = _clientFactory.CreateClient();
                 var response = await client.SendAsync(request);
                 string responseContent = await response.Content.ReadAsStringAsync();
-                _context.Vtex.Logger.Info("ValidateKeyAndToken", null, $"[{response.StatusCode}] {responseContent}");
-                validated = response.IsSuccessStatusCode;
+                _context.Vtex.Logger.Info("ValidateKeyAccessGranted", null, $"[{response.StatusCode}] {responseContent}");
+                keyHasAccess = response.IsSuccessStatusCode && responseContent.Equals("true");
             }
             catch (Exception ex)
             {
-                _context.Vtex.Logger.Error("ValidateKeyAndToken", null, $"Error validating key and token '{key}'", ex);
+                _context.Vtex.Logger.Error("ValidateKeyAccessGranted", null, $"Error validating access for key '{key}'", ex);
             }
 
-            return validated;
+            return keyAndTokenValidated && keyHasAccess;
         }
 
         public async Task<VtexOrder> GetOrderInformation(string orderId)