+++ dotnet/Models/ValidateKeyAndToken.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ReviewsRatings.Models
+{
+    public class ValidateKeyAndToken
+    {
+        public string AppKey { get; set; }
+        public string AppToken { get; set; }
+    }
+}